Fetch SBOM packages with concurrency of 10
