+++ .env
@@ -9,7 +9,7 @@ CODETOTAL_HTTP_PORT=8081
 CODETOTAL_HTTP_HOST=127.0.0.1
 CODETOTAL_WS_PORT=8080
 CODETOTAL_WS_HOST=127.0.0.1
-DEBUG_MODULES=actions,megalinter,stores,transport
+DEBUG_MODULES=actions,megalinter,stores,transport,sbom
 
 # FRONTEND
 CODETOTAL_UPLOAD_FILE_LIMIT_BYTES=10000000
